requirements.txt
# FastAPI and Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI/ML
openai==1.12.0
tavily-python==0.3.3

# Validation
pydantic==2.6.1
pydantic-settings==2.2.1

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
tenacity==8.2.3

# CORS
python-jose==3.3.0
